Added merger
Created merge_files(). It merges all of the document token files, but still keeps them separated by the starting letter of the token (a-z) plus a file for the tokens not starting with a letter. Also lowered() all tokens in build_index() (i stemmed them with PortStemmer in the last commit)